Create Script.py
This script utilizes the `os` module to work with file paths and the `shutil` module to move files. It defines the `organize_files` function that takes the `folder_path` as input. It gets all the files in the specified folder using `os.listdir` and filters out directories using `os.path.isfile`. 

Next, it iterates over each file and extracts the file extension using `os.path.splitext`. It then creates a folder with the extension name (e.g., "TXT_Files" for ".txt" files) by converting the extension to uppercase and appending "_Files". If the folder doesn't exist, it creates one using `os.makedirs`.

Finally, it moves the file into the respective folder using `shutil.move`, specifying the source file path and the destination folder path.

To use this script, you need to provide the `folder_path` variable with the path to the folder you want to organize. Make sure to replace `"C:/Path/To/Folder"` with the actual path to your folder.

When you run the script, it will organize the files in the specified folder based on their extensions, creating separate folders for each file type and moving the files into their corresponding folders.